fix/validation

+++ src/main/java/com/woodongleee/src/userMatch/UserMatchController.java
@@ -13,7 +13,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
-import static com.woodongleee.config.BaseResponseStatus.EMPTY_PARAMETER;
+import static com.woodongleee.config.BaseResponseStatus.*;
 
 @RestController
 @RequestMapping("user-match")
@@ -41,7 +41,12 @@ public class UserMatchController {
                                                                          @RequestParam String endTime){
         try{
             int userIdx = jwtService.getUserIdx();
-            //GetUserByJwtRes getUserByJwtRes = userProvider.getUserByJwt(userIdx);
+            if(userProvider.checkUserExist(userIdx) == 0){
+                throw new BaseException(USER_DOES_NOT_EXIST);
+            }
+            if(userProvider.checkUserStatus(userIdx).equals("INACTIVE")){
+                throw new BaseException(LEAVED_USER);
+            }
 
             List<GetUserMatchPostInfoRes> MatchPosts = userMatchProvider.getUserMatchPosts(userIdx, town, startTime, endTime);
             return new BaseResponse<>(MatchPosts);
@@ -58,7 +63,12 @@ public class UserMatchController {
     public BaseResponse<String> applyUserMatch(@PathVariable("matchPostIdx") int matchPostIdx){
         try{
             int userIdx = jwtService.getUserIdx();
-            //GetUserByJwtRes getUserByJwtRes = userProvider.getUserByJwt(userIdx);
+            if(userProvider.checkUserExist(userIdx) == 0){
+                throw new BaseException(USER_DOES_NOT_EXIST);
+            }
+            if(userProvider.checkUserStatus(userIdx).equals("INACTIVE")){
+                throw new BaseException(LEAVED_USER);
+            }
 
             userMatchService.applyUserMatch(userIdx, matchPostIdx);
             String result = "용병 신청을 성공하였습니다.";
@@ -74,7 +84,12 @@ public class UserMatchController {
     public BaseResponse<String> cancelApplyUserMatch(@PathVariable("matchPostIdx") int matchPostIdx){
         try{
             int userIdx = jwtService.getUserIdx();
-            //GetUserByJwtRes getUserByJwtRes = userProvider.getUserByJwt(userIdx);
+            if(userProvider.checkUserExist(userIdx) == 0){
+                throw new BaseException(USER_DOES_NOT_EXIST);
+            }
+            if(userProvider.checkUserStatus(userIdx).equals("INACTIVE")){
+                throw new BaseException(LEAVED_USER);
+            }
 
             userMatchService.cancelApplyUserMatch(userIdx, matchPostIdx);
             String result = "용병 신청을 취소하였습니다.";
@@ -103,7 +118,12 @@ public class UserMatchController {
             if(userIdx != createUserMatchPostReq.getUserIdx()){
                 return new BaseResponse<>(BaseResponseStatus.INVALID_JWT);
             }
-            //GetUserByJwtRes getUserByJwtRes = userProvider.getUserByJwt(userIdx);
+            if(userProvider.checkUserExist(userIdx) == 0){
+                throw new BaseException(USER_DOES_NOT_EXIST);
+            }
+            if(userProvider.checkUserStatus(userIdx).equals("INACTIVE")){
+                throw new BaseException(LEAVED_USER);
+            }
 
             return userMatchService.createUserMatchPost(userIdx, teamScheduleIdx, createUserMatchPostReq.getContents());
 
@@ -131,7 +151,12 @@ public class UserMatchController {
             if(userIdx != modifyUserMatchPostReq.getUserIdx()){
                 return new BaseResponse<>(BaseResponseStatus.INVALID_JWT);
             }
-            //GetUserByJwtRes getUserByJwtRes = userProvider.getUserByJwt(userIdx);
+            if(userProvider.checkUserExist(userIdx) == 0){
+                throw new BaseException(USER_DOES_NOT_EXIST);
+            }
+            if(userProvider.checkUserStatus(userIdx).equals("INACTIVE")){
+                throw new BaseException(LEAVED_USER);
+            }
 
             return userMatchService.modifyUserMatchPost(userIdx, teamScheduleIdx, modifyUserMatchPostReq.getContents());
 
@@ -147,7 +172,13 @@ public class UserMatchController {
     public BaseResponse<String> deleteUserMatchPost(@PathVariable int teamScheduleIdx){
         try{
             int userIdx = jwtService.getUserIdx();
-            userProvider.checkUserStatus(userIdx);
+            if(userProvider.checkUserExist(userIdx) == 0){
+                throw new BaseException(USER_DOES_NOT_EXIST);
+            }
+            if(userProvider.checkUserStatus(userIdx).equals("INACTIVE")){
+                throw new BaseException(LEAVED_USER);
+            }
+
             userMatchService.deleteUserMatchPost(userIdx, teamScheduleIdx);
             String result = "용병 모집글 삭제를 완료하였습니다.";
             return new BaseResponse<>(result);
@@ -162,6 +193,13 @@ public class UserMatchController {
     public BaseResponse<List<UserMatchApplyInfo>> getUserMatchApplyList(@PathVariable int teamScheduleIdx){
         try{
             int userIdx = jwtService.getUserIdx();
+            if(userProvider.checkUserExist(userIdx) == 0){
+                throw new BaseException(USER_DOES_NOT_EXIST);
+            }
+            if(userProvider.checkUserStatus(userIdx).equals("INACTIVE")){
+                throw new BaseException(LEAVED_USER);
+            }
+
             List<UserMatchApplyInfo> userMatchApplyInfoList = userMatchProvider.getUserMatchApplyList(userIdx, teamScheduleIdx);
             return new BaseResponse<>(userMatchApplyInfoList);
         }
